+++ CHANGELOG.md
@@ -4,6 +4,8 @@ UNRELEASED
 -------------------------
 * Add "serverIPAddress" field to entries.
 * Set bodySize for requests correctly.
+* Set bodySize and compression for responses correctly.
+* Add _transferSize field for responses, just like Chrome does.
 
 version 0.1.0 2017-03-05
 -------------------------